fix: Replace fake() with $this->faker in factories for production compatibility
The fake() helper function is only available in development environments.
Production deployments with APP_ENV=production don't have access to this
global helper, causing 'Call to undefined function fake()' errors.

Changed all factories to use $this->faker instead of fake() which is
available in all environments as it's a property of the Factory base class.

This allows seeders to run in production if needed (though seeders should
generally only be used for development/demo purposes).

Affected factories:
- BuildingFactory
- CompanyFactory
- ConfigurationFileFactory
- DivisionFactory
- GatewayFactory
- LocationFactory
- MeterFactory
- SiteFactory
- UserFactory

Resolves: Call to undefined function Database\Factories\fake()